+++ client/my-sites/stats/stats-subscribers-chart-section/index.tsx
@@ -7,6 +7,7 @@ import Intervals from 'calypso/blocks/stats-navigation/intervals';
 import useSubscribersQuery from 'calypso/my-sites/stats/hooks/use-subscribers-query';
 import StatsModulePlaceholder from '../stats-module/placeholder';
 import StatsPeriodHeader from '../stats-period-header';
+import { hideFractionNumber } from './chart-utils';
 import SubscribersNavigationArrows from './subscribers-navigation-arrows';
 import type uPlot from 'uplot';
 
@@ -47,20 +48,6 @@ function transformData( data: SubscribersData[] ): uPlot.AlignedData {
 
 	return [ x, y ];
 }
-
-// Hide fractional values on the y-axis.
-export const hideFractionNumber = ( self: uPlot, splits: number[] ) => {
-	splits = splits.map( ( split ) => Math.floor( split ) );
-	const newSplits = [ splits[ 0 ] ];
-	for ( let i = 1; i < splits.length; i++ ) {
-		if ( splits[ i ] === splits[ i - 1 ] ) {
-			newSplits.push( null as unknown as number );
-		} else {
-			newSplits.push( splits[ i ] );
-		}
-	}
-	return newSplits;
-};
 
 export default function SubscribersChartSection( {
 	siteId,
